+++ src/lib.rs
@@ -45,9 +45,9 @@ pub use crate::algorithms::paths::{
     is_eulerian, radius,
 };
 pub use crate::algorithms::properties::{
-    assortativity_degree, avg_nearest_neighbor_degree, betweenness, closeness, count_triangles,
-    density, edge_betweenness, eigenvector_centrality, girth, harmonic_centrality, has_loop,
-    has_multiple, is_loop, is_multiple, is_simple, mean_distance, pagerank, reciprocity,
+    assortativity_degree, avg_nearest_neighbor_degree, betweenness, closeness, closeness_weighted,
+    count_triangles, density, edge_betweenness, eigenvector_centrality, girth, harmonic_centrality,
+    has_loop, has_multiple, is_loop, is_multiple, is_simple, mean_distance, pagerank, reciprocity,
     transitivity_local_undirected, transitivity_undirected,
 };
 pub use crate::algorithms::traversal::{BfsTree, bfs, bfs_tree, dfs};